refactor: DTO 네이밍 변경
